fix:added links in README and updated package name (#1320)
