+++ templateview/src/test/java/com/iluwatar/templateview/TemplateViewTest.java
@@ -22,19 +22,38 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwater.templateview;
+package com.iluwatar.templateview;
 
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.mockito.Mockito.*;
 
-/**
- * Application test
- */
-class AppTest {
+class TemplateViewTest {
 
   @Test
-  void shouldExecuteWithoutException() {
-    // Verify that main() method executes without throwing exceptions
-    assertDoesNotThrow(() -> App.main(new String[]{}));
+  void testRenderHomePage() {
+    // Create a spy for HomePageView
+    TemplateView homePage = spy(HomePageView.class);
+
+    // Call the render method
+    homePage.render();
+
+    // Verify that the steps of rendering are executed in the correct order
+    verify(homePage).printHeader(); // Header is printed
+    verify(homePage).renderDynamicContent(); // Dynamic content specific to home page
+    verify(homePage).printFooter(); // Footer is printed
+  }
+
+  @Test
+  void testRenderContactPage() {
+    // Create a spy for ContactPageView
+    TemplateView contactPage = spy(ContactPageView.class);
+
+    // Call the render method
+    contactPage.render();
+
+    // Verify that the steps of rendering are executed in the correct order
+    verify(contactPage).printHeader(); // Header is printed
+    verify(contactPage).renderDynamicContent(); // Dynamic content specific to contact page
+    verify(contactPage).printFooter(); // Footer is printed
   }
 }
